third_party/blink/web_tests/external/wpt/cookies/partitioned-cookies/partitioned-cookies.tentative.https-expected.txt

This is a testharness.js-based test.
PASS Test partitioned cookies
PASS Partitioned cookies accessible on the top-level site they are created in via HTTP
PASS Partitioned cookies accessible on the top-level site they are created in via DOM
PASS Cross-site window opened correctly
FAIL Partitioned cookies are not accessible on a different top-level site via HTTP assert_equals: Expected __Host-pchttp to not be available on a different top-level site expected false but got true
PASS Opening embedded cookie site frame
FAIL Partitioned cookies are not accessible on a different top-level site via DOM assert_equals: Expected __Host-pchttp to not be available on a different top-level site expected false but got true
Harness: the test ran to completion.
